requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.3.0
marshmallow-sqlalchemy==1.4.2
Flask-RESTX==1.3.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.3
pytest-flask==1.3.0
solana==0.30.2
solders==0.18.1
base58==2.1.1

